scripts/requirements.txt
pymupdf>=1.24.0
anthropic>=0.40.0
pytesseract>=0.3.10
pillow>=10.0.0
